SSaP Tone
Tone shaping with a low shelf, a high shelf and a parametric EQ, all on the same frequency.

This is an attempt at replacing the classic amp tone stack using digital filter modules.
The bass and treble controls are shelf filters which can raise or lower by 20dB. The midrange
is a parametric EQ which includes its own Q value. The frequency for all three filters is determined
by a single frequency value. The result is a pedal flexible enough to do subtle tone shaping, wah effects,
, strong midrange scoops, or midrange boosts